Fix DBAAS write error
In Redis 6 backup of DB key-values to RDB file was enabled by default.
In Redis version 5 RDB file backup was disabled.
In RIC we do not have discs attached to DBAAS pods, hence we shouldn't
never try to write RDB backup file as an background operation in Redis
server. So, as a fix disable 'save' and 'append-only-file' features in
Redis by adding below configurations to redis.conf file:
 * save ""
 * appendonly "no"

Issue-Id: RIC-698

Signed-off-by: Timo Tietavainen <timo.tietavainen@nokia.com>
Change-Id: I258226ef00f10c00cccd471fd3e088779443fadb